refactor(backup): move common utils out to backup_utils
