+++ src/common/replication_common.cpp
@@ -588,13 +588,6 @@ void replica_helper::load_meta_servers(/*out*/ std::vector<dsn::rpc_address> &se
     dassert(servers.size() > 0, "no meta server specified in config [%s].%s", section, key);
 }
 
-const std::string cold_backup_constant::APP_METADATA("app_metadata");
-const std::string cold_backup_constant::APP_BACKUP_STATUS("app_backup_status");
-const std::string cold_backup_constant::CURRENT_CHECKPOINT("current_checkpoint");
-const std::string cold_backup_constant::BACKUP_METADATA("backup_metadata");
-const std::string cold_backup_constant::BACKUP_INFO("backup_info");
-const int32_t cold_backup_constant::PROGRESS_FINISHED = 1000;
-
 const std::string backup_restore_constant::FORCE_RESTORE("restore.force_restore");
 const std::string backup_restore_constant::BLOCK_SERVICE_PROVIDER("restore.block_service_provider");
 const std::string backup_restore_constant::CLUSTER_NAME("restore.cluster_name");
@@ -644,126 +637,5 @@ const std::string bulk_load_constant::BULK_LOAD_METADATA("bulk_load_metadata");
 const std::string bulk_load_constant::BULK_LOAD_LOCAL_ROOT_DIR("bulk_load");
 const int32_t bulk_load_constant::PROGRESS_FINISHED = 100;
 
-namespace cold_backup {
-std::string get_policy_path(const std::string &root, const std::string &policy_name)
-{
-    std::stringstream ss;
-    ss << root << "/" << policy_name;
-    return ss.str();
-}
-
-std::string
-get_backup_path(const std::string &root, const std::string &policy_name, int64_t backup_id)
-{
-    std::stringstream ss;
-    ss << get_policy_path(root, policy_name) << "/" << backup_id;
-    return ss.str();
-}
-
-std::string get_app_backup_path(const std::string &root,
-                                const std::string &policy_name,
-                                const std::string &app_name,
-                                int32_t app_id,
-                                int64_t backup_id)
-{
-    std::stringstream ss;
-    ss << get_backup_path(root, policy_name, backup_id) << "/" << app_name << "_" << app_id;
-    return ss.str();
-}
-
-std::string get_replica_backup_path(const std::string &root,
-                                    const std::string &policy_name,
-                                    const std::string &app_name,
-                                    gpid pid,
-                                    int64_t backup_id)
-{
-    std::stringstream ss;
-    ss << get_policy_path(root, policy_name) << "/" << backup_id << "/" << app_name << "_"
-       << pid.get_app_id() << "/" << pid.get_partition_index();
-    return ss.str();
-}
-
-std::string get_app_meta_backup_path(const std::string &root,
-                                     const std::string &policy_name,
-                                     const std::string &app_name,
-                                     int32_t app_id,
-                                     int64_t backup_id)
-{
-    std::stringstream ss;
-    ss << get_policy_path(root, policy_name) << "/" << backup_id << "/" << app_name << "_" << app_id
-       << "/meta";
-    return ss.str();
-}
-
-std::string get_app_metadata_file(const std::string &root,
-                                  const std::string &policy_name,
-                                  const std::string &app_name,
-                                  int32_t app_id,
-                                  int64_t backup_id)
-{
-    std::stringstream ss;
-    ss << get_app_meta_backup_path(root, policy_name, app_name, app_id, backup_id) << "/"
-       << cold_backup_constant::APP_METADATA;
-    return ss.str();
-}
-
-std::string get_app_backup_status_file(const std::string &root,
-                                       const std::string &policy_name,
-                                       const std::string &app_name,
-                                       int32_t app_id,
-                                       int64_t backup_id)
-{
-    std::stringstream ss;
-    ss << get_app_meta_backup_path(root, policy_name, app_name, app_id, backup_id) << "/"
-       << cold_backup_constant::APP_BACKUP_STATUS;
-    return ss.str();
-}
-
-std::string get_current_chkpt_file(const std::string &root,
-                                   const std::string &policy_name,
-                                   const std::string &app_name,
-                                   gpid pid,
-                                   int64_t backup_id)
-{
-    std::stringstream ss;
-    ss << get_replica_backup_path(root, policy_name, app_name, pid, backup_id) << "/"
-       << cold_backup_constant::CURRENT_CHECKPOINT;
-    return ss.str();
-}
-
-std::string get_remote_chkpt_dirname()
-{
-    // here using server address as suffix of remote_chkpt_dirname
-    rpc_address local_address = dsn_primary_address();
-    std::stringstream ss;
-    ss << "chkpt_" << local_address.ipv4_str() << "_" << local_address.port();
-    return ss.str();
-}
-
-std::string get_remote_chkpt_dir(const std::string &root,
-                                 const std::string &policy_name,
-                                 const std::string &app_name,
-                                 gpid pid,
-                                 int64_t backup_id)
-{
-    std::stringstream ss;
-    ss << get_replica_backup_path(root, policy_name, app_name, pid, backup_id) << "/"
-       << get_remote_chkpt_dirname();
-    return ss.str();
-}
-
-std::string get_remote_chkpt_meta_file(const std::string &root,
-                                       const std::string &policy_name,
-                                       const std::string &app_name,
-                                       gpid pid,
-                                       int64_t backup_id)
-{
-    std::stringstream ss;
-    ss << get_remote_chkpt_dir(root, policy_name, app_name, pid, backup_id) << "/"
-       << cold_backup_constant::BACKUP_METADATA;
-    return ss.str();
-}
-
-} // namespace cold_backup
 } // namespace replication
 } // namespace dsn
