+++ src/meta/meta_backup_service.cpp
@@ -1,3 +1,20 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 #include <dsn/utility/filesystem.h>
 #include <dsn/utility/time_utils.h>
 #include <dsn/utility/output_utils.h>
@@ -7,6 +24,7 @@
 #include "meta_service.h"
 #include "server_state.h"
 #include "block_service/block_service_manager.h"
+#include "common/backup_utils.h"
 
 namespace dsn {
 namespace replication {
